Intuitionize section "Opposite ring" (#4572)
* Add opposite rings to iset.mm

This is the oppR syntax and df-oppr .  Copied without change from
set.mm.

* Add opprvalg to iset.mm

This is opprval from set.mm with additional set existence conditions.
The proof is similar to a portion of the set.mm proof.

* Add opprmulfvalg to iset.mm

This is opprmulfval from set.mm with an additional set existence
condition.  The proof needs intuitionizing but is similar to a portion
of the set.mm proof.

* Add opprmulg to iset.mm

This is opprmul from set.mm with set existence conditions added.  The
proof is similar to the set.mm proof.

* Add crngoppr to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof.

* Add opprbasg and oppraddg to iset.mm

This is opprbas and oppradd from set.mm with additional set existence
conditions.  The proofs are similar to the set.mm proofs.

Includes lemma opprsllem which is an intuitionized version of opprlem
from set.mm.

* Add opprring to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a number of
places but is basically the set.mm proof.

* Add opprringbg to iset.mm

This is opprringb from set.mm with an additional set existence
condition.  The proof needs intuitionizing but is based on the set.mm
proof.

* Add opprex to iset.mm

This will be useful in proving some of the other oppR theorems.

* Add mgpex to iset.mm

A set existence theorem for mulGrp .

* Add oppr0g and oppr1g to iset.mm

These are oppr0 and oppr1 from set.mm with additional set
existence conditions.  The proofs needs intuitionizing but are similar to
the set.mm proofs.

* Add opprnegg to iset.mm

This is opprneg from set.mm with an additional set existence condition.
The proof is based on the set.mm proof.

* add opprsubg to mmil.html

* Add mulgpropdg to iset.mm

This is mulgpropd from set.mm but adds set existence conditions and puts two
hypotheses in deduction form.  The proof needs a little intuitionizing
but is basically the set.mm proof.

* Add mulgass3 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.