Intuitionize structure restriction (#4591)
* Add df-iress to iset.mm

This is similar to df-ress from set.mm but only includes the
non-trivial case.  Given reasonable conditions (for example
those in strsetsid ) the two cases give the same results, and
df-ress is infeasible in iset.mm because it wouldn't typically
be easy to show that ( Base ` w ) C_ x is decidable.

Remove df-ress from iset.mm; as described above it would be hard to work
with and would not have any of the advantages it has in set.mm.

- Add ressvalsets to iset.mm

This is similar to ressval in set.mm but adjusted for the difference
between set.mm's df-ress and iset.mm's df-iress .

- Add ressex to iset.mm

This replaces ovex but for |`s

- Add ressval2 to iset.mm

Stated as in set.mm but the proof in iset.mm is trivial from ressvalsets .
(In fact, we retain ressval2 only for consistency with set.mm).

- Add ressval3d to iset.mm

Stated as in set.mm.  The proof is pretty simple because of df-iress
(and in fact does not need the Fun S and E e. dom S hypotheses, but they
are kept for consistency with set.mm).

- Add ressid2 to mmil.html.  Remove from iset.mm.

- Add strressid to iset.mm

This is ressid from set.mm with additional conditions on the structure.
Replaces the ressid which had been in iset.mm which relied on the
previous definition of structure restriction.

* Add ressbasd and ressbas2d to iset.mm

ressbasd is ressbas from set.mm with a W e. X condition added and put into
deduction form.

ressbas2d is ressbas2 from set.mm put into deduction form in a similar way to
ressbas / ressbasd .

* Add ressbasssd to iset.mm

This is ressbasss from set.mm placed into deduction form and with set
existence conditions added.  The proof is similar to a portion of the
set.mm proof.

* Add resseqnbasd to iset.mm

This is resseqnbas from set.mm with some added conditions (similar to
other structure restriction theorems).  The proof is based on a portion
of the set.mm proof.

This replaces the theorem which was called resslem until recently.

* Revise ress0 entry in mmil.html

Give the reason we don't have it (in light of df-iress ).

* Add ressinbasd to iset.mm

This is ressinbasd from set.mm, placed into deduction form and with an
additional W e. V condition.

* WIP ressressg

* Add ressressg to iset.mm

This is ressress from set.mm with an additional set existence condition.
The proof is similar to a portion of the set.mm proof but is also
moderately different especially in details.

* Add ressabsg to iset.mm

This is ressabs from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.