Intuitionize ring units (#4618)
* Add stoic1a and stoic1b to iset.mm

These are similar to con3 and thus hold intuitionistically.

* Add elrelimasn to iset.mm

Stated as in set.mm.  Portions of the proof are based on the set.mm
proof but the need to change relimasn to elimag means that significant
parts of the proof are new.

* Add eliniseg2 to iset.mm

Copied without change from set.mm.

* Remove elreimasng from iset.mm

This was intended as an intuitionization of elrelimasn .
But it turns out that elrelimasn itself can be proved, so we don't need
it (and it was unused).

* Add dvdsrex to iset.mm

This is a theorem that in set.mm would be handled by fvex .
The proof is basd on a portion of the dvdsrvald proof.

* Add isunitd to iset.mm

This is isunit from set.mm put in deduction form and with a R e. SRing
condition added.  The proof needs a fair bit of intuitionizing but is
based on the set.mm proof.

* Add 1unit to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add unitcld to iset.mm

This is unitcl from set.mm put in deduction form with a R e. SRing
condition.  The proof needs some intuitionizing but is basically the
set.mm proof.

* Add unitssd to iset.mm

This is unitss from set.mm put in deduction form with a R e. SRing
condition.  The proof follows from unitcld .

* Add opprunitd to iset.mm

This is opprunit from set.mm put in deduction form with an added R e.
Ring condition.  The proof needs a fair bit of intuitionizing but is
basically the set.mm proof.

* Add crngunit to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy dvdsunit from set.mm to iset.mm

* Add unitmulcl to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in many places but
is basically the set.mm proof.

* Add unitmulclb to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* Add unitgrpbasd to iset.mm

This is unitgrpbas from set.mm put in deduction form and with a R e.
SRing condition added.  The proof is similar to the set.mm proof.

* Add ressplusgd to iset.mm

This is ressplusg from set.mm put in deduction form and with a set
existence condition for G added.  The proof is similar to the set.mm
proof.

* Add unitgrp to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing in most
places but is basically the set.mm proof.

* Add subcmnd to iset.mm

This is subcmn from set.mm put in deduction form and with a set
existence condition added.  The proof is similar to the set.mm proof.

* Add unitabl to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add ringidss to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add unitgrpid to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
a small amount of intuitionization.

* Add issubmnd to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy issubm2 from set.mm to iset.mm

Also, adjust mmil.html accordingly.

* Add unitsubm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Modify unitgrpid comment

This is in set.mm and iset.mm.